Removed working files, moved files between directories, and cleaned up code in final .py file.
